Add PhantomWatcher registration into base PhantomCore
